+++ games_services/example/lib/main.dart
@@ -1,7 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:games_services/games_services.dart';
-import 'package:games_services/models/achievement.dart';
-import 'package:games_services/models/score.dart';
 
 void main() => runApp(MyApp());