+++ games_services/lib/games_services.dart
@@ -1,12 +1,11 @@
 import 'dart:async';
-import 'package:flutter/services.dart';
 import 'package:games_services_platform_interface/game_services_platform_interface.dart';
-import 'package:games_services_platform_interface/helpers.dart';
 import 'package:games_services_platform_interface/models/achievement.dart';
 import 'package:games_services_platform_interface/models/score.dart';
+export 'package:games_services_platform_interface/models/achievement.dart';
+export 'package:games_services_platform_interface/models/score.dart';
 
 class GamesServices {
-  static const MethodChannel _channel = const MethodChannel('games_services');
 
   /// Unlock an [achievement].
   /// [Achievement] takes three parameters:
@@ -15,11 +14,7 @@ class GamesServices {
   /// [percentComplete] the completion percent of the achievement, this parameter is
   /// optional in case of iOS.
   static Future<String> unlock({achievement: Achievement}) async {
-    GamesServicesPlatform.instance.unl
-    return await _channel.invokeMethod("unlock", {
-      "achievementID": achievement.id,
-      "percentComplete": achievement.percentComplete,
-    });
+    return await GamesServicesPlatform.instance.unlock(achievement: achievement);
   }
 
   /// Submit a [score] to specific leader board.
@@ -28,31 +23,23 @@ class GamesServices {
   /// [iOSLeaderboardID] the leader board id that you want to send the score for in case of iOS.
   /// [value] the score.
   static Future<String> submitScore({score: Score}) async {
-    return await _channel.invokeMethod("submitScore", {
-      "leaderboardID": score.leaderboardID,
-      "value": score.value,
-    });
+    return await GamesServicesPlatform.instance.submitScore(score: score);
   }
 
   /// It will open the achievements screen.
   static Future<String> showAchievements() async {
-    return await _channel.invokeMethod("showAchievements");
+    return await GamesServicesPlatform.instance.showAchievements();
   }
 
   /// It will open the leaderboards screen.
   static Future<String> showLeaderboards({iOSLeaderboardID = ""}) async {
-    return await _channel
-        .invokeMethod("showLeaderboards", {"iOSLeaderboardID": iOSLeaderboardID});
+    return await GamesServicesPlatform.instance.showLeaderboards(iOSLeaderboardID: iOSLeaderboardID);
   }
 
   /// To sign in the user.
   /// You need to call the sign in before making any action,
   /// (like sending a score or unlocking an achievement).
   static Future<String> signIn() async {
-    if (Helpers.isPlatformAndroid) {
-      return await _channel.invokeMethod("silentSignIn");
-    } else {
-      return await _channel.invokeMethod("signIn");
-    }
+    return await GamesServicesPlatform.instance.signIn();
   }
 }