+++ games_services_platform_interface/CHANGELOG.md
@@ -1,3 +1,11 @@
 ## 1.0.0
 
 - Initial open-source release.
+
+## 1.0.1
+
+- Remove the static keyword.
+
+## 1.0.2
+
+- Remove the static keyword for showAchievements, showLeaderboards and signIn.